refactor(evaluations): extract consistency-suite fixtures to keep the test under the size rule
